requirements.txt
openai
python-dotenv
requests
Pillow
numpy
opencv-python
